tfmodel/requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.17.0
fastapi==0.67.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
idna==3.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
passlib==1.7.4
Pillow==8.3.1
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.22
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.14.0
Werkzeug==2.0.1
wrapt==1.12.1
